move MethodOption to config package
