+++ global/method_config.go
@@ -17,11 +17,6 @@
 
 package global
 
-import (
-	"strconv"
-	"time"
-)
-
 // MethodConfig defines method config
 type MethodConfig struct {
 	InterfaceId                 string
@@ -37,100 +32,4 @@ type MethodConfig struct {
 	ExecuteLimitRejectedHandler string `yaml:"execute.limit.rejected.handler" json:"execute.limit.rejected.handler,omitempty" property:"execute.limit.rejected.handler"`
 	Sticky                      bool   `yaml:"sticky"   json:"sticky,omitempty" property:"sticky"`
 	RequestTimeout              string `yaml:"timeout"  json:"timeout,omitempty" property:"timeout"`
-}
-
-type MethodOption func(*MethodOptions)
-
-func WithInterfaceId(id string) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.InterfaceId = id
-	}
-}
-
-func WithInterfaceName(name string) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.InterfaceName = name
-	}
-}
-
-func WithName(name string) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.Name = name
-	}
-}
-
-func WithRetries(retries int) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.Retries = strconv.Itoa(retries)
-	}
-}
-
-func WithLoadBalance(lb string) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.LoadBalance = lb
-	}
-}
-
-func WithWeight(weight int64) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.Weight = weight
-	}
-}
-
-func WithTpsLimitInterval(interval int) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.TpsLimitInterval = strconv.Itoa(interval)
-	}
-}
-
-func WithTpsLimitRate(rate int) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.TpsLimitRate = strconv.Itoa(rate)
-	}
-}
-
-func WithTpsLimitStrategy(strategy string) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.TpsLimitStrategy = strategy
-	}
-}
-
-func WithExecuteLimit(limit int) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.ExecuteLimit = strconv.Itoa(limit)
-	}
-}
-
-func WithExecuteLimitRejectedHandler(handler string) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.ExecuteLimitRejectedHandler = handler
-	}
-}
-
-func WithSticky() MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.Sticky = true
-	}
-}
-
-func WithRequestTimeout(millSeconds time.Duration) MethodOption {
-	return func(opts *MethodOptions) {
-		opts.Method.RequestTimeout = millSeconds.String()
-	}
-}
-
-type MethodOptions struct {
-	Method *MethodConfig
-}
-
-func defaultMethodOptions() *MethodOptions {
-	return &MethodOptions{Method: &MethodConfig{}}
-}
-
-func NewMethodOptions(opts ...MethodOption) *MethodOptions {
-	defOpts := defaultMethodOptions()
-	for _, opt := range opts {
-		opt(defOpts)
-	}
-	return defOpts
 }
